+++ backend/public/sql/astro_pdf_gate.sql
@@ -1,0 +1,34 @@
+CREATE DATABASE astro_pdf_db;
+
+
+INSERT INTO pdfs (title, size, type, url, isActive) VALUES
+('UP1.1. Aspectos generales de las arquitecturas web - CE1.a)', '1.197 KB', 'Manual', 'up1-1.pdf', 1),
+('UP1.2. Fundamentos y protocolos de servidores web - CE1.b)', '1.524 KB', 'Manual', 'up1-2.pdf', 1),
+('UP1.3. Estructura y recursos de aplicaciones web - CE1.g)', '818 KB', 'Manual', 'up1-3.pdf', 1),
+('UP1.4. Requisitos del proceso de implementación de aplicaciones web - CE1.h)', '683 KB', 'Manual', 'up1-4.pdf', 1),
+('UP1.5. Instalación y configuración de servidores web - CE1.c)', '889 KB', 'Manual', 'up1-5.pdf', 1),
+('UP1.6. Instalación y configuración de servidores de aplicaciones - CE1.d)', '959 KB', 'Manual', 'up1-6.pdf', 1),
+('UP2.1. Parámetros del servidor web - CE2.a)', '498 KB', 'Manual', 'up2-1.pdf', 1),
+('UP2.2. Activación y configuración de módulos - CE2.b)', '582 KB', 'Manual', 'up2-2.pdf', 1),
+('UP2.3. Creación y configuración de virtual hosts - CE2.c)', '632 KB', 'Manual', 'up2-3.pdf', 1),
+('UP2.4. Configuración de mecanismos de autenticación y control de acceso...', '985 KB', 'Manual', 'up2-4.pdf', 1),
+('UP2.5. Obtención e instalación de certificados digitales - CE2.e)', '610 KB', 'Manual', 'up2-5.pdf', 1),
+('UP2.6. Mecanismos de seguridad en las comunicaciones web - CE2.f)', '601 KB', 'Manual', 'up2-6.pdf', 1),
+('UP2.7. Herramientas de generación de logs - CE2.j)', '529 KB', 'Manual', 'up2-7.pdf', 1),
+('UP3.1. Servicios del servidor de aplicaciones - CE3.a)', '513 KB', 'Manual', 'up3-1.pdf', 1),
+('UP3.2. Archivos de configuración y bibliotecas compartidas - CE3.b)', '700 KB', 'Ficha de Datos', 'up3-2.pdf', 1),
+('UP3.3. Despliegue de la aplicación web en servidor de aplicaciones - CE3.c)', '594 KB', 'Ficha de Datos', 'up3-3.pdf', 1),
+('UP3.4. Mecanismos de seguridad del servidor de aplicaciones - CE3.d)', '508 KB', 'Ficha de Datos', 'up3-4.pdf', 1),
+('UP3.5. Componentes web del servidor de aplicaciones - CE3.e)', '701 KB', 'Ficha de Datos', 'up3-5.pdf', 1),
+('UP4.1. Herramientas para generar documentación - CE6.a)', '566 KB', 'Ficha de Datos', 'up4-1.pdf', 1),
+('UP4.2. Documentación de PHP - CE6.b)', '699 KB', 'Ficha de Datos', 'up4-2.pdf', 1),
+('UP4.3. Documentación del proyecto - CE6.c) y CE6.d)', '1.644 KB', 'Ficha de Datos', 'up4-3.pdf', 1),
+('UP4.4. Control de versiones con Git - CE6.e) y CE6.g)', '2.075 KB', 'Ficha de Datos', 'up4-4.pdf', 1),
+('UP4.5. Accesibilidad y seguridad del control de versiones - CE6.f)', '569 KB', 'Ficha de Datos', 'up4-5.pdf', 1),
+('UP4.6. Repositorio GitHub - CE6.h)', '1.666 KB', 'Ficha de Datos', 'up4-6.pdf', 1),
+('UP5.2. Creación de usuarios para acceso remoto al servidor FTP - CE4.b)', '619 KB', 'Ficha de Datos', 'up5-2.pdf', 1),
+('UP5.3. Acceso al servidor en modo activo y pasivo - CE4.c)', '531 KB', 'Ficha de Datos', 'up5-3.pdf', 1),
+('UP5.4. Pruebas de cliente FTP - CE4.d)', '697 KB', 'Ficha de Datos', 'up5-4.pdf', 1),
+('UP5.5. Instalación, configuración y conexión de SFTP - CE4.e)', '616 KB', 'Ficha de Datos', 'up5-5.pdf', 1),
+('UP5.6. Servicios de transferencia de archivos integrados en servidores web', '763 KB', 'Ficha de Datos', 'up5-6.pdf', 1);
+