+++ backend/src/main.ts
@@ -1,8 +1,12 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
+import { NestExpressApplication } from '@nestjs/platform-express';
+import { join } from 'path';
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
+  const app = await NestFactory.create<NestExpressApplication>(AppModule);
+  // Configurar la carpeta estática
+  app.useStaticAssets(join(__dirname, '..', 'public'));
 
   // ACTIVAR CORS: Esto permite que Astro le hable a Nest
   app.enableCors({